+++ src/WriteMAP.cpp
@@ -57,7 +57,7 @@ int libsiedler2::loader::WriteMAP(const std::string& file, const ArchivInfo& ite
                 continue;
             if(items.get(i)->getBobType() == BOBTYPE_MAP)
             {
-                nr = i;
+                nr = static_cast<long>(i);
                 break;
             }
